reconfigure and restart controller in more cases
a controller should be reconfigured if

* node units depart;
* a backup controller departs.

Moreover, as there is a two-way exchange with node units, relying on
slurm.conf changes on a controller is not a reliable indicator of a
need for restart - nodes are not present in sinfo after the model
settles.

Restarting on every changed event is more expensive, however, this could
be optimized via goal-state hook tool execution and observing "joining"
node units.