Manual edit of cluster config include file now triggers reconfiguration
If the active Slurm controller detects that the file
/etc/slurm-llnl/slurm-<clustername>.conf has changed between the
update-status hooks, reconfiguration of the controller is triggered.

The controller reconfiguration reads the slurm-<clustername>.conf and
includes it in the controller_conf dict, so that nodes and backup
controllers may consume it and get an identical Slurm configuration.

Cosmic and disco as been added as valid series. Disco has Slurm 18,
which is the latest stable version.